traffic_signs_recognizer/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
grpcio>=1.50.0
protobuf>=4.0.0
